refactor(server): modify vad,asr file struct and vad cache to vad factory
