+++ apps/server/api/src/config/vad.rs
@@ -2,20 +2,20 @@ use serde::Deserialize;
 
 #[derive(Debug, Default, Deserialize, Clone)]
 pub struct VadConfig {
-    model: Option<String>,
+    path: Option<String>,
     num_threads: Option<i32>,
 }
 
 impl VadConfig {
     pub fn new() -> Self {
         Self {
-            model: Some(String::from("data/vad/model.onnx")),
+            path: Some(String::from("data/vad/model/silero/")),
             num_threads: Some(4),
         }
     }
 
-    pub fn model(&self) -> &str {
-        self.model.as_deref().unwrap_or_default()
+    pub fn path(&self) -> &str {
+        self.path.as_deref().unwrap_or_default()
     }
 
     pub fn num_threads(&self) -> i32 {
