+++ apps/server/api/src/vad/model/mod.rs
@@ -1,0 +1,1 @@
+pub mod silero;
